Initial Sphinx documentation.
Currently this only contains the auto-generated API docs, but over time we can
extend and improve this to include additional documentation and tutorials that
are not in docstrings.

To build and view the docs run the following and navigate to
http://localhost:8888:

    $ cd docs/
    $ pip install -r requirements.txt
    $ cd _build/html
    $ python -m http.server 8888

We will automate the process of generating and pushing the docs to an
appropriate destination (e.g. gh-pages and/or readthedocs) when commits are
merged or releases are tagged.

PiperOrigin-RevId: 249470120
Change-Id: I5d0de111a19320e9c52e5d08aa6e395aeff68022