Set `module.__test__` for bazel doctest.
We run two lots of doctest, first via bazel to provide quick feedback when
testing iteratively. Secondly we run doctest on all the generated Sphinx docs
via `test.sh` to ensure samples in our documentation run as described.

When running via bazel we need to set `__test__` appropriately so doctest runs
over imported symbols. This is not needed for the Sphinx tests since Sphinx runs
all code samples that are included in the docs.

PiperOrigin-RevId: 255849190
Change-Id: I697c248fa1c198b7fae201e7d12a3572a7a55d82